chore(extensions): update VS Code extensions chore(dependencies): update file paths in dependency scripts chore(environment): update file paths in environment setup script chore(setup): update file paths in setup script docs(developers): update repository name and issue links chore(codeserver): update default workspace path in .env.example
Update the list of VS Code extensions in the extensions.json file. Update the file paths in the dependency scripts, environment setup script, and setup script to use the correct directory name. Update the repository name and issue links in the developers documentation. Update the default workspace path in the .env.example file for the codeserver container.

fix(docker-compose): change image name to uppercase
fix(Makefile): change application name to uppercase
fix(setupLinuxMachine.sh): change file paths to uppercase
fix(updateHosts.sh): change file paths to uppercase
fix(dockerInstall.sh): change file paths to uppercase
fix(dockerUpdate.sh): change file paths to uppercase
fix(provisionDockerService.sh): change file paths to uppercase
fix(provisionDockerService.sh): change file paths to uppercase
fix(selectService.sh): change file paths to uppercase

Change the image name in the docker-compose file to uppercase 'Wolflith'. Change the application name in the Makefile to uppercase 'Wolflith'. Change the file paths in the setupLinuxMachine.sh, updateHosts.sh, dockerInstall.sh, dockerUpdate.sh, provisionDockerService.sh, and selectService.sh scripts to uppercase.

fix(docker): fix path to PersonalizationFunc.sh in setupService.sh
fix(proxmox): fix path to Ansible playbooks in provisionLxcDockerService.sh
fix(proxmox): fix path to PersonalizationFunc.sh in provisionLxc.sh
fix(tools): fix path to Ansible playbooks in executeCommand.sh

The changes were made to fix the paths to the required files in the scripts. The paths were incorrect and needed to be updated to the correct locations.

fix(PCSMenu): fix file paths in source commands
feat(PCSMenu): update aliases to use Wolflith instead of wolflith
fix(PCSInstall): fix folder name in ifolder variable
feat(PCSInstall): update folder name to Wolflith
fix(PCSMenu): fix BASE_DIR variable to use Wolflith instead of wolflith
fix(PCSUpdate): fix directory name in cd command
docs: update repository name to Wolflith in README.md
Update file paths in the source commands to fix a bug in the PCSMenu and PCSFunc.sh files. Update the aliases in the PCSInstall.sh file to use the new folder name 'Wolflith'. Fix the folder name in the ifolder variable in the PCSInstall.sh file. Update the BASE_DIR variable in the PCSMenu.sh file to use the new folder name 'Wolflith'. Fix the directory name in the cd command in the PCSUpdate.sh file. Update the repository name to 'Wolflith' in the README.md file.

chore(vault): update module path and dependencies
Update the module path and dependencies in the go.mod file for the Vault script. Replace the module path with "github.com/NorkzYT/Wolflith/Scripts/Vault" and add/update the required dependencies.

chore(vault): update dependencies
Update the dependencies in the go.sum file for the Vault project. This includes adding new versions for the following packages:
- github.com/davecgh/go-spew v1.1.1
- github.com/fatih/color v1.7.0
- github.com/go-test/deep v1.0.2
- github.com/google/go-cmp v0.5.7
- github.com/hashicorp/go-hclog v0.16.2
- github.com/mattn/go-colorable v0.1.6
- github.com/mattn/go-isatty v0.0.12
- github.com/pmezard/go-difflib v1.0.0

This update is done to ensure that the project is using the latest versions of these dependencies, which may include bug fixes or new features.

chore(deps): update testify/testify to v1.7.0 and add gopkg.in/yaml.v3 v3.0.0-20200313102051-9f266ea9e77c
Update the dependency testify/testify to version 1.7.0. Also, add a new dependency gopkg.in/yaml.v3 v3.0.0-20200313102051-9f266ea9e77c. These updates and additions are done to keep the project up to date with the latest versions and to utilize new features and bug fixes provided by the dependencies.

feat(vault): add main.go and remove vault-pull.go
Add a new file 'main.go' to the Scripts/Vault directory. This file contains the main function that handles pulling or pushing secrets to Hashicorp Vault based on the specified action. The 'action' flag is used to determine the action to be performed.

Remove the file 'vault-pull.go' from the Scripts/Vault directory. This file is no longer needed as the functionality has been moved to the 'main.go' file.

These changes are done to improve the organization and structure of the Vault scripts and to provide a more streamlined way of pulling and pushing secrets to Hashicorp Vault.

feat(vault): move common functions to separate package
fix(vault): fix error message for missing environment variables

Move the common functions from the 'vault-push.go' file to a new package called 'vaultcommon' in the 'vault_common.go' file. This improves code organization and reusability.

Fix the error message in the 'InitVaultClient' function to provide a more informative message when one or more required environment variables are not set.

feat(vault): add vault_pull_module.go and vault_push_module.go
fix(package.json): fix package name capitalization

Add two new files, `vault_pull_module.go` and `vault_push_module.go`, to the `Scripts/Vault/vaultPull` and `Scripts/Vault/vaultPush` directories respectively. These files contain functions for pulling secrets from Vault and pushing secrets to Vault.

Fix the capitalization of the package name in the `package.json` file. Change `"name": "wolflith"` to `"name": "Wolflith"`.
