+++ Docker/Both/codeserver/.env.example
@@ -2,5 +2,5 @@
 #OPENVSCODE_SERVER_CONNECTION_SECRET=${OPENVSCODE_SERVER_CONNECTION_SECRET}
 OPENVSCODE_SERVER_SUDO_PASSWORD='xxx'
 #OPENVSCODE_SERVER_SUDO_PASSWORD_HASH=${OPENVSCODE_SERVER_SUDO_PASSWORD_HASH}
-OPENVSCODE_SERVER_DEFAULT_WORKSPACE=/config/workspace/wolflith.code-workspace
+OPENVSCODE_SERVER_DEFAULT_WORKSPACE=/config/workspace/Wolflith.code-workspace
 TZ=America/New_York
