+++ PCSMenu/Functions/Ansible/Setup Linux Machine/setupLinuxMachine.sh
@@ -5,9 +5,9 @@
 # License : General Public License GPL-3.0-or-later
 ######################################################################
 
-source /opt/wolflith/PCSMenu/PCSFunc.sh
+source /opt/Wolflith/PCSMenu/PCSFunc.sh
 
-HOSTS_FILE="/opt/wolflith/Ansible/inventory/hosts.yaml"
+HOSTS_FILE="/opt/Wolflith/Ansible/inventory/hosts.yaml"
 
 ## ----------------------------------------------------------------------------------------------------------- ##
 
@@ -106,7 +106,7 @@ execute_ansible_playbooks() {
     run_playbook() {
         local playbook=$1
         blueprint "Running playbook: $playbook"
-        if ! ansible-playbook "/opt/wolflith/Ansible/playbooks/$playbook.yml" -i "$HOSTS_FILE" $LIMIT_OPTION; then
+        if ! ansible-playbook "/opt/Wolflith/Ansible/playbooks/$playbook.yml" -i "$HOSTS_FILE" $LIMIT_OPTION; then
             redprint "Failed to execute playbook: $playbook"
             return 1
         fi
