+++ PCSMenu/Functions/Docker/Docker Update/dockerUpdate.sh
@@ -30,7 +30,7 @@ function docker_update() {
 
             cyanprint "Executing Playbook... (Please be patient, may take more than 10 minutes to complete)"
             # Execute the Ansible playbook for the specified target(s), capturing output
-            if ! output=$(ansible-playbook /opt/wolflith/Ansible/playbooks/docker-update.yml -i "/opt/wolflith/Ansible/inventory/hosts.yaml" -l "$ansible_playbook_targets" 2>&1); then
+            if ! output=$(ansible-playbook /opt/Wolflith/Ansible/playbooks/docker-update.yml -i "/opt/Wolflith/Ansible/inventory/hosts.yaml" -l "$ansible_playbook_targets" 2>&1); then
                 redprint "An error occurred during playbook execution:"
                 echo "$output" # Display the captured error output
