+++ PCSMenu/Functions/Proxmox/Scripts/provisionLxc.sh
@@ -5,7 +5,7 @@
 # License : General Public License GPL-3.0-or-later
 ######################################################################
 
-source /opt/wolflith/PCSMenu/PersonalizationFunc.sh
+source /opt/Wolflith/PCSMenu/PersonalizationFunc.sh
 
 # Default values
 default_vCPU=2
@@ -17,11 +17,11 @@ default_api_user="root"
 default_node="pve"
 
 SELECTED_SERVICE=$(cat /tmp/selected_docker_service.txt)
-HOSTS_FILE="/opt/wolflith/Ansible/inventory/hosts.yaml"
+HOSTS_FILE="/opt/Wolflith/Ansible/inventory/hosts.yaml"
 TARGET_HOST="$ANSIBLE_PLAYBOOK_TARGET"
 
 # Extract ansible_host (WAN IP), ansible_user (API user), and ansible_ssh_pass (API password)
-read api_host api_user api_password <<<$(awk -v target="$TARGET_HOST" '
+read -r api_host api_user api_password <<<$(awk -v target="$TARGET_HOST" '
 BEGIN {
     RS=""; FS="\n"
     host_found = 0; user = ""; pass = ""
@@ -64,7 +64,7 @@ blueprint "Configuration for Proxmox LXC Provisioning"
 # If api_host is empty, fallback to user input
 if [ -z "$api_host" ]; then
     greenprint "Enter Proxmox VE API host (e.g., 24.156.99.202 [WAN IP]): "
-    read api_host
+    read -r api_host
 else
     greenprint "Proxmox VE API host (WAN IP) extracted: $api_host"
 fi
@@ -72,7 +72,7 @@ fi
 # Prompt for API User if not extracted
 if [ -z "$api_user" ]; then
     greenprint "Enter Proxmox VE API user (default: $default_api_user): "
-    read api_user
+    read -r api_user
     api_user=${api_user:-$default_api_user}
     echo ""
 else
@@ -82,25 +82,25 @@ fi
 # Prompt for API password if not extracted
 if [ -z "$api_password" ]; then
     greenprint "Enter Proxmox VE API password: "
-    read -sp "" api_password
+    read -rsp "" api_password
     echo ""
 else
     greenprint "Proxmox VE API password extracted"
 fi
 
 greenprint "Enter Proxmox VE node name (default: $default_node): "
-read node
+read -r node
 node=${node:-$default_node}
 
 greenprint "Use default specifications? (vCPU: $default_vCPU, RAM: $default_RAM MB, lxc_storage_amount: $default_Storage GB) [Y/n] "
-read use_defaults
+read -r use_defaults
 if [[ "$use_defaults" =~ ^[Nn] ]]; then
     greenprint "Enter number of vCPU: "
-    read vCPU
+    read -r vCPU
     greenprint "Enter RAM size in MB: "
-    read RAM
+    read -r RAM
     greenprint "Enter storage size in GB: "
-    read lxc_storage_amount
+    read -r lxc_storage_amount
 else
     vCPU=$default_vCPU
     RAM=$default_RAM
@@ -108,11 +108,11 @@ else
 fi
 
 greenprint "Enter storage name for VM disk (default: $default_storage_name): "
-read lxc_storage_name
+read -r lxc_storage_name
 lxc_storage_name=${lxc_storage_name:-$default_storage_name}
 
 greenprint "Enter the name of the storage that contains CT Templates (default: $default_ct_storage): "
-read ct_storage
+read -r ct_storage
 ct_storage=${ct_storage:-$default_ct_storage}
 
 # Generate and display root password for LXC
@@ -127,7 +127,7 @@ greenprint "LXC root password for $hostname saved to $PASSWORDS_FILE"
 # Additional questions based on the Proxmox Ansible docs
 cyanprint "The default hostname will be the service you selected to install: $SELECTED_SERVICE"
 greenprint "Would you like to define a custom hostname suffix or leave empty to use default? [Enter suffix or leave blank]: "
-read hostname_suffix
+read -r hostname_suffix
 if [ -z "$hostname_suffix" ]; then
     hostname="${SELECTED_SERVICE}"
 else
@@ -138,12 +138,12 @@ hostname=$(echo "$hostname" | tr '[:upper:]' '[:lower:]') # Convert to lowercase
 cyanprint "Network interface configuration:"
 ip_config="dhcp"
 greenprint "Use DHCP or static IP? [dhcp/static]: "
-read ip_config_choice
+read -r ip_config_choice
 if [[ "$ip_config_choice" == "static" ]]; then
     greenprint "Enter static IP address (e.g., 192.168.1.100/24): "
-    read ip_address
+    read -r ip_address
     greenprint "Enter gateway IP address (e.g., 192.168.0.1): "
-    read gw_ip
+    read -r gw_ip
     ip_config="{'net0':'name=eth0,gw=$gw_ip,ip=$ip_address,bridge=vmbr0'}"
 else
     ip_config="{'net0':'name=eth0,ip=dhcp,bridge=vmbr0'}"
