+++ README.md
@@ -11,15 +11,15 @@
 
 </div>
 
-At its core, wolflith harnesses the power of Ansible and Bash to offer an unparalleled level of control and flexibility, making it the backbone for deploying and managing your technological services. This Ansible-Bash-based toolkit simplifies complex tasks, ensuring that even those new to homelabs can achieve professional-grade setups. With wolflith, you're not just building a server; you're crafting a fully customized, scalable digital environment tailored to your needs. It's the perfect blend of simplicity and sophistication for the modern tech enthusiast.
+At its core, Wolflith harnesses the power of Ansible and Bash to offer an unparalleled level of control and flexibility, making it the backbone for deploying and managing your technological services. This Ansible-Bash-based toolkit simplifies complex tasks, ensuring that even those new to homelabs can achieve professional-grade setups. With Wolflith, you're not just building a server; you're crafting a fully customized, scalable digital environment tailored to your needs. It's the perfect blend of simplicity and sophistication for the modern tech enthusiast.
 
 Note: This repository is in BETA and still a work in progress, I welcome contributions to expand its capabilities and usability.
 
 ## How to Install
 
 ```bash
-mkdir -p /opt/wolflith
-wget -qO /opt/PCSInstall.sh https://raw.githubusercontent.com/NorkzYT/wolflith/main/PCSMenu/PCSInstall.sh
+mkdir -p /opt/Wolflith
+wget -qO /opt/PCSInstall.sh https://raw.githubusercontent.com/NorkzYT/Wolflith/main/PCSMenu/PCSInstall.sh
 chmod +x /opt/PCSInstall.sh
 sudo /opt/PCSInstall.sh
 ```
@@ -207,6 +207,6 @@ Created by NorkzYT with 💛
 - [**Temp**](Temp)
   - [**Auto**](Temp/Auto)
   - [**Bin**](Temp/Bin)
-- [Developing wolflith](DEVELOPERS.md)
+- [Developing Wolflith](DEVELOPERS.md)
 
 <!-- tree generated by markdown-notes-tree ends here -->
