+++ package.json
@@ -1,5 +1,5 @@
 {
-	"name": "wolflith",
+	"name": "Wolflith",
 	"description": "Streamline Your Tech Stack Deployment with One Command.",
 	"version": "1.0.0",
 	"author": "NorkzYT",
